Move gitea into separate package
